requirements.txt
psycopg2-binary==2.9.10
flask==3.0.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
python-dotenv==1.0.0
tabulate==0.9.0